refactor(tests): move helpers to separate file
