+++ test/PluginManager.test.js
@@ -2,61 +2,15 @@ import PluginManager from '../src/PluginManager.js';
 import { Plugin } from '@chammy/plugin-helper';
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import { join } from 'path';
+import { createParameterTest } from './helpers/helpers.js';
+import { createMethodTest } from './helpers/helpers.js';
+import { createLoadTest } from './helpers/helpers.js';
+import { createFindPathTest } from './helpers/helpers.js';
+import { getPath } from './helpers/helpers.js';
 
 const expect = chai.expect;
 
 chai.use( chaiAsPromised );
-
-function createParameterTest( { invalid = [], test = () => {}, errorType = TypeError, errorMsg = '' } = {} ) {
-	return () => {
-		invalid.forEach( ( param ) => {
-			expect( () => {
-				test( param );
-			} ).to.throw( errorType, errorMsg );
-		} );
-	};
-}
-
-function createMethodTest( object, method ) {
-	return () => {
-		expect( object ).to.have.property( method );
-		expect( object[ method ] ).to.be.a( 'function' );
-	};
-}
-
-function createLoadTest( paths, pluginsPropertyValue ) {
-	return () => {
-		const pluginManager = new PluginManager();
-
-		return pluginManager.load( paths ).then( ( plugins ) => {
-			expect( plugins ).to.be.an( 'array' );
-			expect( plugins ).to.have.lengthOf( paths.length );
-
-			plugins.forEach( ( plugin ) => {
-				expect( Reflect.getPrototypeOf( plugin ) ).to.equal( Plugin );
-			} );
-
-			expect( [ ...pluginManager.plugins ] ).to.deep.equal( pluginsPropertyValue || plugins );
-		} );
-	};
-}
-
-function createFindPathTest( { patterns = '', path = undefined, expected = [] } = {} ) {
-	return () => {
-		const pluginManager = new PluginManager();
-
-		return pluginManager.find( patterns, path ).then( ( paths ) => {
-			expect( paths ).to.be.an( 'array' );
-			expect( paths ).to.have.lengthOf( expected.length );
-			expect( paths ).to.have.members( expected );
-		} );
-	};
-}
-
-function getPath( name ) {
-	return `${ join(  __dirname, name ) }`;
-}
 
 describe( 'PluginManager', () => {
 	it( 'is a class', () => {
